Consolidates IARRC color detectors into one node with performance improvements.
