+++ iarrc/CMakeLists.txt
@@ -45,9 +45,8 @@ include_directories(
 )
 
 add_subdirectory(src/stoplight_watcher)
-add_subdirectory(src/color_detector_dragrace)
-add_subdirectory(src/color_detector_circuit)
 add_subdirectory(src/steerer_dragrace)
 add_subdirectory(src/steerer_circuit)
 add_subdirectory(src/navigation_controller)
 add_subdirectory(src/finish_line_watcher)
+add_subdirectory(src/color_detector)
