--- iarrc/launch/color_detector_circuit.launch
@@ -1,4 +1,0 @@
-<launch>
-	<node name="color_detector_circuit" pkg="iarrc" type="color_detector_circuit" output="screen">
-	</node>
-</launch>
