+++ iarrc/src/color_detector/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_executable(color_detector color_detector.cpp)
+target_link_libraries(color_detector ${catkin_LIBRARIES} ${OpenCV_LIBRARIES})
