--- iarrc/src/color_detector_dragrace/CMakeLists.txt
@@ -1,2 +1,0 @@
-add_executable(color_detector_dragrace color_detector_dragrace.cpp)
-target_link_libraries(color_detector_dragrace ${catkin_LIBRARIES} ${OpenCV_LIBRARIES})
